Removed hc.gov-specific customizations to html examples
The extent of the customizations seems to be to move the error messaging below the component. It is actually out of date and doubles all the error icons. This change should actually be addressed in the new doc site, so I'm removing it, especially since hc.gov is moving towards only using the single-input date field.