This is a failed attempt at moving away from a query param for theme
Besides being a lot more files to build, there are a lot of unsolved problems with this method of appending `/[theme-name]/` at the end of URL paths. I don't like it, and it doesn't work very well. I'm concluding that the little UX improvement of not seeing a flash of core-themed content on first load is not worth the trouble.