Another little test for using our react components in custom elements
Was curious how the styles work. If we can somehow match the styles in the parent document in our shadow dom, we can get our components styled with no extra bytes over the wire...however, I don't know if there are performance hits to loading all the styles into a shadow dom.