This fixes cursor jump for controlled label masks
Before this change, it wasn't actually using that stored `currentValue`. That meant uncontrolled components never specified a `value`, so they weren't susceptible to the defect, and controlled components didn't get the benefit of avoiding the re-rendering of the `input` element, which is what I hypothesized was causing the issue to begin with. In my research, I found [some talk about this general problem with controlled React inputs on StackOverflow](https://stackoverflow.com/questions/46000544/react-controlled-input-cursor-jumps) and in [this blog post](https://dev.to/kwirke/solving-caret-jumping-in-react-inputs-36ic), but I don't think they're understanding the problem correctly. It's a little fuzzy for me too, and I'm following hunches based on a weak mental model.